Minor changes to problems 2 and 3

+++ 3. Factorial.cpp
@@ -5,7 +5,7 @@ int main() {
 	int n;
 	cin >> n;
 
-	int ans = 1;
+	long long int ans = 1;
 	for (int i = 1; i < n + 1; i++) {
 		ans = ans * i;
 	}
